wind.rose(openair)	R Documentation

Traditional wind rose plot

Description

Traditional wind rose plot that plots wind speed intervals by 30 degree angles.

Usage


wind.rose(polar, ws.int = 2, type = "year", cols = "default", use.all = FALSE, main = "", auto.text = TRUE, ...)

Arguments

polar 	 A data frame containing fiels ws and ws{wd}

ws.int 	 Wind speed interval. Default is 1 m/s but for low met masts with low mean wind speeds a value of 0.5 m/s may be better.

type 	 Default is year, can also be month.

cols 	 Colours to be used for plotting. Options include default, increment, heat, jet and user defined. For user defined the user can supply a list of colour names recognised by R (type colours() to see the full list). An example would be cols = c("yellow", "green", "blue", "black")

use.all 	 If set to TRUE it will produce a single plot using all data.

main 	 Title of plot.

auto.text 	 Either TRUE (default) or FALSE. If TRUE titles and axis labels will automatically try and format pollutant names and units properly e.g. by subscripting the `2' in NO2.

... 	 Other graphical parameters to pass to wind.rose.

Author(s)

David Carslaw

See Also

polar.freq for a more flexible version that considers other statistics and pollutant concentrations.

Examples


# load example data from package
data(mydata)

# basic plot
wind.rose(mydata)

# wind rose using all data
wind.rose(mydata, use.all = TRUE)

# plot wind roses by month of the year
wind.rose(mydata, type = "month")

[Package openair version 1.0 Index]